[#1960] Organize wire-related components into separate module
