+++ integrations/zenoh/gateway-backend/src/wire/mod.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2025 Contributors to the Eclipse Foundation
+// Copyright (c) 2026 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -10,9 +10,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-pub mod event;
-mod factory;
-pub mod publish_subscribe;
-mod wake_handler;
-
-pub use factory::*;
+pub mod description;
+pub mod descriptor;
+pub mod fingerprint;
+pub mod keys;
+pub mod message;
